Moved the healthcheck endpoints to a dedicated file
- Created /src/routes/healthcheck.ts with both / and /details endpoints
- Updated /src/index.ts to use the new healthcheck router
- Removed unused imports from the main index file

The endpoints will work the same:
- GET /healthcheck returns "OK"
- GET /healthcheck/details returns detailed service status
